Update testing and CI

+++ .test.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
 
-set -e
-echo "" > coverage.txt
-
 for d in $(go list ./...); do
-    go test -race -v -coverprofile=profile.out -covermode=atomic $d
-    if [ -f profile.out ]; then
-        cat profile.out >> coverage.txt
-        rm profile.out
+    if [ $1 == "race" ]; then
+        echo "race"
+        go test -race || exit 2
+    else 
+        set -e;
+        echo "" > coverage.txt;
+
+        go test -v -coverprofile=profile.out -covermode=atomic $d || exit 1
+        if [ -f profile.out ]; then
+            cat profile.out >> coverage.txt;
+            rm profile.out;
+        fi
     fi
 done
